ProvBase/Calix: Modem class chosen via configfile, not separate select.
Up to this point there was – as a proof-of-concept – a dropdown to
choose the correct modem type (= qualified_model_class). After that
configfile/qos for that class where selectable.
After a discussion in the team we decided to bring back the old logic
(modem type is not selected expicitely but implicitely by choosing the
configfile). Background: We don't want our customers to adapt to the new
logic.

But still the correct class needs to be instantiated (e.g. CalixOnt
instead of Modem) and qualified_model_class needs to be set correctly.
This has been solved by redirecting and/or injecting data.
As this effects GUI and API create and update functionality (which
mostly is handled generic) there may be unforseen problems related to
this. If you have trouble related to Modem and derived classes this
commit is a good starting point…